# Use fast PRNG to generate `_id` for documents created outside methods