--- a/History.md
+++ b/History.md
@@ -7,6 +7,13 @@ message for more details.
 
 ## v.NEXT
 
+* `coll.insert()` now uses a faster (but cryptographically insecure)
+  algorithm to generate document IDs when called outside of a method
+  and an `_id` field is not explicitly passed. With this change, there
+  are no longer two algorithms used to generate document
+  IDs. `Random.id()` can still be used to generate crpytographically
+  secure document IDs. #5161
+
 * No longer include the `json` package by default, which contains code for
   `JSON.parse` and `JSON.stringify`.  (The last browser to not support JSON
   natively was Internet Explorer 7.)