--- a/packages/mongo/collection.js
+++ b/packages/mongo/collection.js
@@ -62,14 +62,18 @@ Mongo.Collection = function (name, options) {
   switch (options.idGeneration) {
   case 'MONGO':
     self._makeNewID = function () {
-      var src = name ? DDP.randomStream('/collection/' + name) : Random;
+      var src = name
+            ? DDP.randomStream('/collection/' + name)
+            : Random.fast;
       return new Mongo.ObjectID(src.hexString(24));
     };
     break;
   case 'STRING':
   default:
     self._makeNewID = function () {
-      var src = name ? DDP.randomStream('/collection/' + name) : Random;
+      var src = name
+            ? DDP.randomStream('/collection/' + name)
+            : Random.fast;
       return src.id();
     };
     break;