--- a/packages/random/package.js
+++ b/packages/random/package.js
@@ -5,6 +5,7 @@ Package.describe({
 
 Package.onUse(function (api) {
   api.use('underscore');
+  api.use('ecmascript');
   api.export('Random');
   api.addFiles('random.js');
   api.addFiles('deprecated.js');
@@ -12,6 +13,7 @@ Package.onUse(function (api) {
 
 Package.onTest(function(api) {
   api.use('random');
+  api.use('ecmascript');
   api.use('tinytest');
   api.addFiles('random_tests.js', ['client', 'server']);
 });