@glasser 2015-09-15
Sure.

Initial impression: call it `Random.insecure` instead of `Random.fast`.  There's a reason it's not the default, so we should give it a name that emphasizes why you should think carefully about using it instead of why you should be excited about using it.