@glasser 2015-09-15
This requires a doc or history.md note to make clear that anyone relying on the behavior for security needs the explicit Random.id() now.