@avital 2015-09-15
Thanks @glasser, I'll make those changes. Did you go through the entire change or should I expect potentially more comments?