@avital 2015-09-20
and added History.md note. doesn't look like the docs need changing.

@glasser i think this is ready to be merged