@avital 2015-10-09
The problem is that Meteor doesn't ship with an algorithm that's both
cryptographically secure and can be seeded (which is necessary for latency
compensated methods that generate multiple documents)

On Fri, Oct 9, 2015 at 3:48 AM, Mitar <notifications@github.com> wrote:

> How does one make so that secure IDs are generated also in the simulation
> mode?
>
> Can secure IDs be enabled for the whole collection with some options to
> the collection constructor? (Like when we can choose between string and
> ObjectID IDs? Maybe there could also be "secure string" ID options there.)
>
> —
> Reply to this email directly or view it on GitHub
> <https://github.com/meteor/meteor/pull/5161#issuecomment-146831838>.
>