@mitar 2015-10-09
Hm, I thought that this was how it was before?

Is there an extension point where community could then develop packages for this?