@avital 2015-10-09
Yes Mitar that's right.

On Fri, Oct 9, 2015 at 2:25 PM, Mitar <notifications@github.com> wrote:

> Oh, I get it. The text in history doc was a bit hard to understand. :-)
> OK. So this deals only with IDs outside of a method call, which are now
> made in the same way as inside a method call. So it does not change the
> behavior for method calls and situation when you want the same ID between
> client and server.
>
> —
> Reply to this email directly or view it on GitHub
> <https://github.com/meteor/meteor/pull/5161#issuecomment-146989314>.
>